pyproject.toml
[project]
name = "info-to-action-agent"
version = "0.1.0"
description = "Information-to-Action Agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "google-api-python-client>=2.188.0",
    "google-auth>=2.47.0",
    "google-auth-oauthlib>=1.2.4",
    "google-genai>=1.60.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "rich>=13.7.0",
    "typer>=0.21.1",
    "youtube-transcript-api>=1.2.3",
]
